prototype computed properties
Add a hook to View, testing whether a mapping exposes a member function isComputed(DatumCoord) and if the value is true for a given DatumCoord on access. If true, the compute member function of the mapping is invoked instead of getBlobNrAndOffset(), allowing the mapping to provide a computed value instead of a memory location.