add bufferguard example
This example showcases a custom LLAMA mapping splitting a 2D array of 3 component vectors into 9 regions.
These 9 regions are the center part of the array, 4 guard regions at the left, right, top and bottom border and the 4 corners of the array.
The content of the blobs belonging to one guard region of a view are copied to blobs of a corresponding guard region of another view.
Such a workflow is typical for HPC applications where the guard regions are exchanged between compute nodes.